bot/requirements.txt
discord.py==2.6.3
ephem==4.2 
pyephem==9.99